+++ client/src/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { render } from 'react-dom';
-import App from './App';
+import Router from './Router';
 
 render(
-  <App />,
+  <Router />,
   document.getElementById('app'),
 );
